web/server/requirements.txt
owslib
pillow
mercantile
# web server
flask
gunicorn
# yolo 8
ultralytics